+++ python/llm/src/ipex_llm/transformers/convert.py
@@ -1188,14 +1188,14 @@ def convert_forward(m, target_m, new_forward):
         convert_forward(sub_m, target_m, new_forward)
 
 
-def convert_hybrid_forward(m, target_m, new_forward, new_device=None):
+def convert_forward_to_xpu(m, target_m, new_forward):
+    # print(m.__class__.__name__)
     if m.__class__ == target_m:
         bound_method = new_forward.__get__(m, m.__class__)
         setattr(m, "forward", bound_method)
-        if new_device is not None:
-            m = m.to(device=new_device)
+        m = m.to(device="xpu", dtype=torch.float16)
     for _, sub_m in m.named_children():
-        convert_hybrid_forward(sub_m, target_m, new_forward, new_device)
+        convert_forward_to_xpu(sub_m, target_m, new_forward)
 
 
 def replace_RotaryEmbed(m, target_m,  replace_embed):
@@ -2018,9 +2018,14 @@ def _optimize_post(model):
     elif model.config.model_type == "deepseek_v3":
         modeling_module_name = model.__class__.__module__
         module = importlib.import_module(modeling_module_name)
-        from ipex_llm.transformers.models.deepseek_v3 import hybrid_DeepseekV3Attention_forward
+        from ipex_llm.transformers.models.deepseek_v3 import (
+            hybrid_DeepseekV3Attention_forward,
+            hybrid_DeepseekV3MLP_forward,
+        )
 
-        convert_hybrid_forward(model, module.DeepseekV3Attention, hybrid_DeepseekV3Attention_forward, "xpu")
+        convert_forward_to_xpu(model, module.DeepseekV3Attention, hybrid_DeepseekV3Attention_forward)
+        convert_forward_to_xpu(model.model.layers[:3], module.DeepseekV3MLP, hybrid_DeepseekV3MLP_forward)
+
     elif model.config.model_type == "baichuan_m1":
         modeling_module_name = model.__class__.__module__
         module = importlib.import_module(modeling_module_name)
